:zap: security with id implemented

+++ app/Http/Controllers/Api/CommentaryController.php
@@ -35,6 +35,9 @@ class CommentaryController extends Controller
      */
     public function store(Request $request)
     {
+        $request->merge([
+            'user_id' => auth()->user()->id,
+        ]);
         $commentary = Commentary::create($request->all());
         ExperienceController::giveExperience($commentary->user_id, 4);
         return response()->json($commentary);
@@ -62,6 +65,9 @@ class CommentaryController extends Controller
      */
     public function update(Request $request, Commentary $commentary)
     {
+        if (auth()->user()->id != $commentary->user_id) {
+            return response()->json(['error' => __('lang.unauthorized')], 401);
+        }
         $commentary->update($request->all());
         return response()->json($commentary);
     }
@@ -74,6 +80,9 @@ class CommentaryController extends Controller
      */
     public function destroy(Commentary $commentary)
     {
+        if (auth()->user()->id != $commentary->user_id) {
+            return response()->json(['error' => __('lang.unauthorized')], 401);
+        }
         $commentary->delete();
         return response()->json($commentary);
     }
